chore: restructure files
